--- data/sub-control01/func/sub-control01_task-music_run-3_events.tsv
@@ -1,19 +1,0 @@
-onset	duration	trial_type
-0	33	tones
-33	3	response
-36	31.5	negative_music
-67.5	3	response
-70.5	31.5	tones
-102	3	response
-105	31.5	positive_music
-136.5	3	response
-139.5	31.5	tones
-171	3	response
-174	31.5	negative_music
-205.5	3	response
-208.5	31.5	tones
-240	3	response
-243	31.5	positive_music
-274.5	3	response
-277.5	31.5	tones
-309	3	response
